Slight reorg of the RouteModule types
