+++ .changeset/slimy-cows-begin.md
@@ -1,0 +1,5 @@
+---
+"react-router": patch
+---
+
+Internal reorg of the `RouteModule` interfaces for deduplication and documentation purposes
