+++ packages/react-router/index.ts
@@ -200,10 +200,13 @@ export type {
   ClientActionFunctionArgs,
   ClientLoaderFunction,
   ClientLoaderFunctionArgs,
+  HeadersArgs,
+  HeadersFunction,
   MetaArgs,
   MetaDescriptor,
   MetaFunction,
   LinksFunction,
+  ServerRouteModule,
 } from "./lib/dom/ssr/routeModules";
 export type { ServerRouterProps } from "./lib/dom/ssr/server";
 export { ServerRouter } from "./lib/dom/ssr/server";
@@ -250,11 +253,6 @@ export type {
   HtmlLinkDescriptor,
   LinkDescriptor,
 } from "./lib/router/links";
-
-export type {
-  HeadersArgs,
-  HeadersFunction,
-} from "./lib/server-runtime/routeModules";
 
 export type { RequestHandler } from "./lib/server-runtime/server";
