+++ packages/react-router/lib/server-runtime/entry.ts
@@ -1,14 +1,14 @@
+import type { RouteModules } from "../dom/ssr/routeModules";
 import type { ServerRouteManifest } from "./routes";
-import type { RouteModules, EntryRouteModule } from "./routeModules";
 
 export function createEntryRouteModules(
   manifest: ServerRouteManifest
-): RouteModules<EntryRouteModule> {
+): RouteModules {
   return Object.keys(manifest).reduce((memo, routeId) => {
     let route = manifest[routeId];
     if (route) {
       memo[routeId] = route.module;
     }
     return memo;
-  }, {} as RouteModules<EntryRouteModule>);
+  }, {} as RouteModules);
 }
