--- packages/react-router/lib/server-runtime/routeModules.ts
@@ -1,56 +1,0 @@
-import type { ActionFunction, LoaderFunction } from "../router/utils";
-import type {
-  ClientActionFunction,
-  ClientLoaderFunction,
-  LinksFunction,
-  MetaFunction,
-} from "../dom/ssr/routeModules";
-
-export interface RouteModules<RouteModule> {
-  [routeId: string]: RouteModule | undefined;
-}
-
-export type HeadersArgs = {
-  loaderHeaders: Headers;
-  parentHeaders: Headers;
-  actionHeaders: Headers;
-  errorHeaders: Headers | undefined;
-};
-
-/**
- * A function that returns HTTP headers to be used for a route. These headers
- * will be merged with (and take precedence over) headers from parent routes.
- */
-export interface HeadersFunction {
-  (args: HeadersArgs): Headers | HeadersInit;
-}
-
-type LdJsonObject = { [Key in string]: LdJsonValue } & {
-  [Key in string]?: LdJsonValue | undefined;
-};
-type LdJsonArray = LdJsonValue[] | readonly LdJsonValue[];
-type LdJsonPrimitive = string | number | boolean | null;
-type LdJsonValue = LdJsonPrimitive | LdJsonObject | LdJsonArray;
-
-/**
- * An arbitrary object that is associated with a route.
- */
-export type RouteHandle = unknown;
-
-export interface EntryRouteModule {
-  clientAction?: ClientActionFunction;
-  clientLoader?: ClientLoaderFunction;
-  ErrorBoundary?: any; // Weakly typed because server-runtime is not React-aware
-  HydrateFallback?: any; // Weakly typed because server-runtime is not React-aware
-  Layout?: any; // Weakly typed because server-runtime is not React-aware
-  default: any; // Weakly typed because server-runtime is not React-aware
-  handle?: RouteHandle;
-  links?: LinksFunction;
-  meta?: MetaFunction;
-}
-
-export interface ServerRouteModule extends EntryRouteModule {
-  action?: ActionFunction;
-  headers?: HeadersFunction | { [name: string]: string };
-  loader?: LoaderFunction;
-}
